api/requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.1.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
